Fix comments appears on refreshing browser
