+++ src/module/comment.js
@@ -10,7 +10,7 @@ const addnewComent = async (id, name, msg) => {
     },
   });
   // eslint-disable-next-line no-use-before-define
-  const arr = getComments(id);
+  const arr = await getComments(id);
   const dataCard = document.querySelector('.comment-card');
   // eslint-disable-next-line no-use-before-define
   displayComents(arr, dataCard);
@@ -26,7 +26,7 @@ const newComment = (id, name, msg) => {
 const CommentForm = (newcommetId, node) => {
   const commentTitle = document.createElement('div');
   commentTitle.classList.add('form-container');
-  commentTitle.innerHTML = '<h3> Add a Comment </h3>';
+  commentTitle.innerHTML = '<h4> Add a Comment </h4>';
   const form = document.createElement('form');
   form.classList.add('form-content');
   form.innerHTML = `<input type="text" class="username" placeholder="Your name" required >
@@ -35,14 +35,29 @@ const CommentForm = (newcommetId, node) => {
   const btncomment = form.querySelector('.btncomment');
   const username = form.querySelector('.username');
   const msg = form.querySelector('.msg');
-  btncomment.addEventListener('click', () => newComment(newcommetId, username, msg));
+  btncomment.addEventListener('click', (e) => {
+    e.preventDefault();
+    newComment(newcommetId, username, msg);
+  });
   commentTitle.appendChild(form);
   node.appendChild(commentTitle);
+};
+
+// counter for number of comments for a single item
+const communtCounter = (comment) => {
+  let counter = comment.length;
+  if (comment.error) {
+    counter = 0;
+  }
+  return counter;
 };
 
 // function to Display comments given for a single item
 const displayComents = (data, node) => {
   node.innerHTML = '';
+  const head = document.createElement('h4');
+  head.innerHTML = `Comments (${communtCounter(data)})`;
+  node.appendChild(head);
   const commentitem = document.createElement('div');
   commentitem.classList.add('comment-items');
   if (!data.error) {
@@ -66,4 +81,4 @@ const getComments = async (id) => {
   return comment;
   // console.log(comment);
 };
-module.exports = { CommentForm, displayComents, getComments };
+module.exports = { communtCounter, CommentForm, displayComents, getComments };
